refactor: added use-team-state hook
